backend/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-cors-headers==3.13.0
django-rest-framework==0.1.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
mysqlclient==2.1.1
PyMySQL==1.0.2
python-decouple==3.6
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.5